Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-invalidnonce.html

<!DOCTYPE html>
<html>
<head>
<script src='../resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['no', 'script-src \'n', 'resources/script.js', ''],
    ['no', 'script-src \'nonce', 'resources/script.js', ''],
    ['no', 'script-src \'nonce-\'', 'resources/script.js', ''],
    ['no', 'script-src \'nonce-', 'resources/script.js', ''],
    ['no', 'script-src nonce-abcd', 'resources/script.js', ''],
    ['no', 'script-src \'nonce- \'', 'resources/script.js', ''],
    ['no', 'script-src \'nonce-     \'', 'resources/script.js', ''],
    ['no', 'script-src \'nonce- nonces have no spaces\'', 'resources/script.js', ''],
    ['no', 'script-src \'nonce-{}\'', 'resources/script.js', '{}'],
    ['no', 'script-src \'nonce-/\\\'', 'resources/script.js', '/\\'],
];
</script>
</head>
<body onload="testExperimentalPolicy()">
  <p>
      None of these scripts should execute, as all the nonces are invalid.
  </p>
